Revert "Add ability to deploy from ACR / Docker Hub to ACI (#2265)"
This reverts commit de261fefbc63de660bacc9233d45a1c63200427f.
