+++ src/tree/containers/ContainersTreeItem.ts
@@ -121,9 +121,9 @@ export class ContainersTreeItem extends LocalRootTreeItemBase<DockerContainer, C
     }
 }
 
-export const NonComposeGroupName = localize('vscode-docker.tree.containers.otherContainers', 'Other Containers');
+const NonComposeGroupName = localize('vscode-docker.tree.containers.otherContainers', 'Other Containers');
 
-export function getComposeProjectName(container: DockerContainer): string {
+function getComposeProjectName(container: DockerContainer): string {
     const labels = Object.keys(container.Labels)
         .map(label => ({ label: label, value: container.Labels[label] }));
